pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydro_replace"
version = "0.1.0"
description = "Pipeline for analyzing baryonic correction models using halo replacement"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["cosmology", "simulations", "baryon correction", "weak lensing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Astronomy",
]

dependencies = [
    "numpy>=1.20",
    "scipy>=1.7",
    "h5py>=3.0",
    "astropy>=5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mpi = ["mpi4py>=3.0"]
power = ["Pylians3"]
bcm = ["BaryonForge>=1.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.0",
    "numpydoc>=1.5",
]
all = [
    "mpi4py>=3.0",
    "Pylians3",
    "BaryonForge>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/hydro_replace"
Documentation = "https://hydro-replace.readthedocs.io"
Repository = "https://github.com/yourusername/hydro_replace"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
